Move proposalsConfig to protocol-configuration folder
