+++ proposals/README.md
@@ -23,7 +23,7 @@ These permissiones are validated against on-chain state in the last e2e test
 ## Step 3: Proposal Mocking and Integration Test
 Write a script following the template of `proposals/dao/fip_x.js`. See below for descriptions of each of the `deploy`, `setup`,`teardown`, and `validate` functions. Only `validate` is required.
 
-Add an object with the key `fip_x` to `end-to-end/proposals_config.ts`, 
+Add an object with the key `fip_x` to `end-to-end/proposalsConfig.ts`, 
 
 Your proposal will be run before any integration tests via `npm run test:e2e`. Fill in the following parameters:
 * deploy - set to true only if you added a deploy script for your proposal in the optional step, otherwise false. This will run your deploy script before the integration tests and add the contract objects as keys in `contracts` parameter of each of the hooks.
@@ -32,7 +32,7 @@ Your proposal will be run before any integration tests via `npm run test:e2e`. F
 ### Step 3a (Optional): deploy() - Contract Deployments:
 Whether a deployment of a new instance of a pre-exisiting or new contract, if your proposal requires a new contract deployment you'll need to write a deploy script.
 
-The deploy script is automatically run before any e2e tests if the deploy flag is set to true in the `end-to-end/proposals_config.ts`. The contract objects will be present in the setup, run, teardown and validate hooks as keys in the `contracts` parameter.
+The deploy script is automatically run before any e2e tests if the deploy flag is set to true in the `end-to-end/proposalsConfig.ts`. The contract objects will be present in the setup, run, teardown and validate hooks as keys in the `contracts` parameter.
 
 This is useful for fully testing the deploy script against a mainnet fork before deploying to mainnet.
 
@@ -65,7 +65,7 @@ Run `DEPLOY_FILE=fip_x npm run deploy:fip`
 
 Run your deploy script if you had one from step 2. Update `/protocol-configuration/mainnetAddresses.ts` with the new contract addresses. 
 
-Update the fork block inside the hardhat config and set the deploy flag to false in the config entry for `fip_x` in `end-to-end/proposals_config.ts`
+Update the fork block inside the hardhat config and set the deploy flag to false in the config entry for `fip_x` in `end-to-end/proposalsConfig.ts`
 
 Finally rerun `npm run test:e2e` and make sure everything passes as expected.
