+++ test/integration/tests/votium-bribe.ts
@@ -3,7 +3,7 @@ import { ethers } from 'hardhat';
 import { NamedContracts } from '@custom-types/types';
 import { getImpersonatedSigner, time } from '@test/helpers';
 import { TestEndtoEndCoordinator } from '@test/integration/setup';
-import proposals from '@test/integration/proposals_config';
+import proposals from '@test/integration/proposalsConfig';
 import { expectRevert } from '@test/helpers';
 import { forceEth } from '../setup/utils';
