+++ tsconfig.json
@@ -24,7 +24,7 @@
     "test/*", "test/**/*",
     "scripts/*", "scripts/**/*",
     "proposals/*", "proposals/**/*",
-    "./types/contracts/*", "./types/contracts/**/*",
+    "./types/contracts/*", "./types/contracts/**/*", "protocol-configuration/proposalsConfig.ts",
   ],
   "exclude": [
     "scripts/deploy/old/*",
